add error type to test
