+++ src/graphql/errors.rs
@@ -7,3 +7,18 @@ pub enum UserError {
     #[error("User with emai `{0}` already exists")]
     UserWithEmailAlreadyExists(String),
 }
+
+#[derive(Error, Debug)]
+pub enum AuthorizationError {
+    #[error(
+        "Required permission is {} but user has permission {}",
+        required,
+        permission
+    )]
+    InsufficientPermission {
+        required: String,
+        permission: String,
+    },
+    #[error("Unable to decode JWT: {0}")]
+    DecodingError(String),
+}
